+++ pages/example/components.config.js
@@ -2,7 +2,6 @@ export default [{
     groupName: '基础组件',
     groupName_en: 'Basic components',
     list: [
-        
     {
         path: '/pages/componentsB/color/color',
         icon: 'color',